+++ rust/operator-binary/src/backend/tls/mod.rs
@@ -28,7 +28,7 @@ use stackable_operator::{
 use time::OffsetDateTime;
 
 use crate::{
-    crd,
+    crd::{self, CertificateKeyGeneration},
     format::{well_known, SecretData, WellKnownSecretData},
     utils::iterator_try_concat_bytes,
 };
@@ -132,12 +132,13 @@ impl SecretBackendError for Error {
 pub struct TlsGenerate {
     ca_manager: ca::Manager,
     max_cert_lifetime: Duration,
+    key_generation: CertificateKeyGeneration,
 }
 
 impl TlsGenerate {
     /// Check if a signing CA has already been instantiated in a specified Kubernetes secret - if
     /// one is found the key is loaded and used for signing certs.
-    /// If no current authority can be found, a new keypair and self signed certificate is created
+    /// If no current authority can be found, a new key pair and self signed certificate is created
     /// and stored for future use.
     /// This allows users to provide their own CA files, but also enables secret-operator to generate
     /// an independent self-signed CA.
@@ -147,6 +148,7 @@ impl TlsGenerate {
             secret: ca_secret,
             auto_generate: auto_generate_ca,
             ca_certificate_lifetime,
+            key_generation,
         }: &crd::AutoTlsCa,
         max_cert_lifetime: Duration,
     ) -> Result<Self> {
@@ -158,11 +160,13 @@ impl TlsGenerate {
                     manage_ca: *auto_generate_ca,
                     ca_certificate_lifetime: *ca_certificate_lifetime,
                     rotate_if_ca_expires_before: Some(*ca_certificate_lifetime / 2),
+                    key_generation: key_generation.clone(),
                 },
             )
             .await
             .context(LoadCaSnafu)?,
             max_cert_lifetime,
+            key_generation: key_generation.clone(),
         })
     }
 }
@@ -171,7 +175,7 @@ impl TlsGenerate {
 impl SecretBackend for TlsGenerate {
     type Error = Error;
 
-    /// Generate a keypair and sign it with the CA key.
+    /// Generate a key pair and sign it with the CA key.
     /// Then add the ca certificate and return these files for provisioning to the volume.
     async fn get_secret_data(
         &self,
@@ -233,7 +237,12 @@ impl SecretBackend for TlsGenerate {
         }
 
         let conf = Conf::new(ConfMethod::default()).unwrap();
-        let pod_key = Rsa::generate(2048)
+
+        let pod_key_length = match self.key_generation {
+            CertificateKeyGeneration::Rsa { length } => length,
+        };
+
+        let pod_key = Rsa::generate(pod_key_length)
             .and_then(PKey::try_from)
             .context(GenerateKeySnafu)?;
         let mut addresses = Vec::new();