+++ rust/operator-binary/src/crd.rs
@@ -4,7 +4,7 @@ use serde::{Deserialize, Serialize};
 use snafu::Snafu;
 use stackable_operator::{
     kube::CustomResource,
-    schemars::{self, JsonSchema},
+    schemars::{self, schema::Schema, JsonSchema},
     time::Duration,
 };
 use stackable_secret_operator_crd_utils::SecretReference;
@@ -44,7 +44,7 @@ pub enum SecretClassBackend {
     /// issues a TLS certificate signed by the Secret Operator.
     /// The certificate authority can be provided by the administrator, or managed automatically by the Secret Operator.
     ///
-    /// A new certificate and keypair will be generated and signed for each Pod, keys or certificates are never reused.
+    /// A new certificate and key pair will be generated and signed for each Pod, keys or certificates are never reused.
     AutoTls(AutoTlsBackend),
 
     /// The [`experimentalCertManager` backend][1] injects a TLS certificate issued
@@ -123,11 +123,71 @@ pub struct AutoTlsCa {
     /// If `autoGenerate: false` then the Secret Operator will log a warning instead.
     #[serde(default = "AutoTlsCa::default_ca_certificate_lifetime")]
     pub ca_certificate_lifetime: Duration,
+
+    /// The algorithm used to generate a key pair and required configuration settings.
+    /// Currently only RSA and a key length of 2048, 3072 or 4096 bits can be configured.
+    #[serde(default)]
+    pub key_generation: CertificateKeyGeneration,
 }
 
 impl AutoTlsCa {
     fn default_ca_certificate_lifetime() -> Duration {
         backend::tls::DEFAULT_CA_CERT_LIFETIME
+    }
+}
+
+#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, JsonSchema)]
+#[serde(rename_all = "camelCase")]
+pub enum CertificateKeyGeneration {
+    Rsa {
+        /// The amount of bits used for generating the RSA keypair.
+        /// Currently, `2048`, `3072` and `4096` are supported. Defaults to `2048` bits.
+        #[schemars(schema_with = "CertificateKeyGeneration::tls_key_length_schema")]
+        length: u32,
+    },
+}
+
+impl CertificateKeyGeneration {
+    pub const RSA_KEY_LENGTH_2048: u32 = 2048;
+    pub const RSA_KEY_LENGTH_3072: u32 = 3072;
+    pub const RSA_KEY_LENGTH_4096: u32 = 4096;
+
+    // Could not get a "standard" enum with assigned values/discriminants to work as integers in the schema
+    // The following was generated and requires the length to be provided as string (we want an integer)
+    // keyGeneration:
+    //   default:
+    //     rsa:
+    //       length: '2048'
+    //   oneOf:
+    //     - required:
+    //         - rsa
+    //   properties:
+    //     rsa:
+    //       properties:
+    //         length:
+    //           enum:
+    //             - '2048'
+    //             - '3072'
+    //             - '4096'
+    //           type: string
+    pub fn tls_key_length_schema(_: &mut schemars::gen::SchemaGenerator) -> Schema {
+        serde_json::from_value(serde_json::json!({
+            "type": "integer",
+            "enum": [
+                Self::RSA_KEY_LENGTH_2048,
+                Self::RSA_KEY_LENGTH_3072,
+                Self::RSA_KEY_LENGTH_4096
+            ]
+        }))
+        .expect("Failed to parse JSON of custom tls key length schema")
+    }
+}
+
+impl Default for CertificateKeyGeneration {
+    fn default() -> Self {
+        Self::Rsa {
+            length: Self::RSA_KEY_LENGTH_2048,
+        }
     }
 }
 
@@ -368,6 +428,9 @@ mod test {
                 secret:
                   name: secret-provisioner-tls-ca
                   namespace: default
+                keyGeneration:
+                  rsa:
+                    length: 3072
         "#;
         let deserializer = serde_yaml::Deserializer::from_str(input);
         let secret_class: SecretClass =
@@ -383,6 +446,9 @@ mod test {
                         },
                         auto_generate: false,
                         ca_certificate_lifetime: DEFAULT_CA_CERT_LIFETIME,
+                        key_generation: CertificateKeyGeneration::Rsa {
+                            length: CertificateKeyGeneration::RSA_KEY_LENGTH_3072
+                        }
                     },
                     max_certificate_lifetime: DEFAULT_MAX_CERT_LIFETIME,
                 })
@@ -418,7 +484,8 @@ mod test {
                             namespace: "default".to_string(),
                         },
                         auto_generate: true,
-                        ca_certificate_lifetime: Duration::from_days_unchecked(100)
+                        ca_certificate_lifetime: Duration::from_days_unchecked(100),
+                        key_generation: CertificateKeyGeneration::default()
                     },
                     max_certificate_lifetime: Duration::from_days_unchecked(31),
                 })